RMX1901: Add keymaster 4.0
Fixes:
W libc    : Unable to set property "ctl.interface_start" to "android.hardware.keymaster@4.0::IKeymasterDevice/default": error code: 0x20
I hwservicemanager: Tried to start android.hardware.keymaster@4.0::IKeymasterDevice/default as a lazy service, but was unable to. Usually this happens when a service is not installed, but if the service is intended to be used as a lazy service, then it may be configured incorrectly.